+++ text/src/diff_util/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) mod change;
+pub(crate) mod common;
+pub(crate) mod constants;
+pub(crate) mod dir_data;
+pub(crate) mod dir_diff;
+pub(crate) mod file_data;
+pub(crate) mod file_diff;
+pub(crate) mod functions;
+pub(crate) mod hunks;
